feat(cleanup): Make cleanup script developer friendly

+++ scripts/cleanup.sh
@@ -6,8 +6,36 @@ LINE="__________________________________________________________________"
 
 echo "${LINE} Starting docker images cleanup..."
 set -Eeuo pipefail
-docker kill $(docker ps -q)
-docker rm $(docker ps -a -q)
-docker rmi $(docker images -q) -f
+# kills = $(docker ps -q | wc -c)
+if [[ $(docker ps -q | wc -c) -ne 0 ]]; then
+    echo "Killing docker containers..."
+    docker kill $(docker ps -q)
+    echo "Done."
+else
+    echo "No Docker containers found to kill off."
+fi
+
+if [[ $(docker ps -a -q | wc -c) -ne 0 ]]; then
+    echo "Removing Docker containers..."
+    docker kill $(docker ps -aq)
+    echo "Done."
+else
+    echo "No Docker containers found to remove."
+fi
+
+if [[ $(docker images -q | wc -c) -ne 0 ]]; then
+    echo "Removing Docker images..."
+    docker kill $(docker images -q)
+    echo "Done."
+else
+    echo "No Docker images found to remove."
+fi
+
+# if [(docker ps -q)]
+# then 
+# docker kill $(docker ps -q)
+# fi
+# docker rm $(docker ps -a -q)
+# docker rmi $(docker images -q) -f
 
 echo "${LINE} Cleanup is complete!"
